WORKAROUND: avoid circular dependency between jupyter_collaboration and jupyter_ydoc
