+++ packages/docprovider/src/chat/index.ts
@@ -6,3 +6,4 @@
 export * from './factory';
 export * from './model';
 export * from './widget';
+export * from './ychat';
